+++ Questionable.Model/Questing/DutyOptions.cs
@@ -6,5 +6,6 @@ public class DutyOptions
     public bool Enabled { get; set; }
     public uint ContentFinderConditionId { get; set; }
     public bool LowPriority { get; set; }
+    public bool? CanUnsync { get; set; }
     public List<string> Notes { get; set; } = [];
 }