+++ Questionable/Configuration.cs
@@ -301,6 +301,8 @@ internal sealed class Configuration : IPluginConfiguration
         public bool OpenEditor { get; set; }
         public bool NamazuPreferCraft { get; set; }
         public bool Debug { get; set; }
+        public bool AutoRedeemRewardItems { get; set; } = true;
+        public HashSet<uint> AutoRedeemItemBlacklist { get; set; } = [];
     }
 
     internal sealed class PriorityConfiguration