+++ Questionable/Model/ItemReward.cs
@@ -11,7 +11,9 @@ public enum EItemRewardType
     Minion,
     OrchestrionRoll,
     TripleTriadCard,
-    FashionAccessory
+    FashionAccessory,
+    Coffer,
+    UnlockLink,
 }
 
 public sealed class ItemRewardDetails(Item item, ElementId elementId)
@@ -29,8 +31,14 @@ public abstract record ItemReward(ItemRewardDetails Item)
     public ElementId ElementId => Item.ElementId;
     public TimeSpan CastTime => Item.CastTime;
     public abstract EItemRewardType Type { get; }
+    internal static bool IsValidCoffer(Item item) =>
+        item.ItemAction.RowId is 1085 or 388 or 367 && item.ItemUICategory.RowId is 61;
+
     internal static ItemReward? CreateFromItem(Item item, ElementId elementId)
     {
+        if (IsValidCoffer(item))
+            return new CofferReward(new(item, elementId));
+
         if (item.ItemAction.Value is { } itemAction &&
             itemAction.Action.Value is { } action)
         {
@@ -48,7 +56,11 @@ public abstract record ItemReward(ItemRewardDetails Item)
 
             if (action.RowId is 3357)
                 return new TripleTriadCardReward(new(item, elementId), (ushort)item.AdditionalData.RowId);
+
+            if (action.RowId is 2633)
+                return new UnlockLinkReward(new(item, elementId), (ushort)item.ItemAction.Value.Data[0]);
         }
+
         return null;
     }
     public abstract bool IsUnlocked();
@@ -93,4 +105,18 @@ public sealed record FashionAccessoryReward(ItemRewardDetails Item, uint Accesso
     public override EItemRewardType Type => EItemRewardType.FashionAccessory;
 
     public override unsafe bool IsUnlocked() => PlayerState.Instance()->IsOrnamentUnlocked(AccessoryId);
+}
+
+public sealed record CofferReward(ItemRewardDetails Item) : ItemReward(Item)
+{
+    public override EItemRewardType Type => EItemRewardType.Coffer;
+
+    public override bool IsUnlocked() => false;
+}
+
+public sealed record UnlockLinkReward(ItemRewardDetails Item, ushort UnlockLinkId) : ItemReward(Item)
+{
+    public override EItemRewardType Type => EItemRewardType.UnlockLink;
+
+    public override unsafe bool IsUnlocked() => UIState.Instance()->IsUnlockLinkUnlocked(UnlockLinkId);
 }