+++ CHANGELOG.md
@@ -1,1 +1,4 @@
-- Feature: Dragon Quest crossover quest (untested, fate spawn times too long) -alydev
+- Feature: We can now exclude duties from the "run content unsynced" option if they cannot be done unsync -Kage, alydev
+- Feature: QST now opens coffers and redeems items automatically. Please report any issues. -Kage, alydev
+  - condition check -Kage
+  - Add Blacklist and option to turn on/off the function - Kage